Conflicts:
	3rdParty/glew/Findglew.cmake
	3rdParty/glfw/Findglfw.cmake
	3rdParty/include/3rdparty/GLFW.h
	3rdParty/opengl/Findglext.cmake
	cmake/kinevox.cmake
	cmake/kinevoxconf.cmake
	cmake/kinevoxconf.h.in

+++ lib/renderer/unittest/GLRendererTest.h
@@ -80,12 +80,12 @@ void Test()
 	{
 		std::stringstream s;
 		s<<renderer.GetViewPort().GetHeight()<<"=="<<Config::Instance().GetInt("DEFAULT_VIEWPORT_HEIGHT");
-		AssertMsg_Test(renderer.GetViewPort().GetHeight()==Config::Instance().GetInt("DEFAULT_VIEWPORT_HEIGHT"), s.str().c_str());
+		AssertMsg_Test(renderer.GetViewPort().GetHeight()==(size_t)Config::Instance().GetInt("DEFAULT_VIEWPORT_HEIGHT"), s.str().c_str());
 	}
 	{
 		std::stringstream s;
 		s<<renderer.GetViewPort().GetWidth()<<"=="<<Config::Instance().GetInt("DEFAULT_VIEWPORT_WIDTH");
-		AssertMsg_Test(renderer.GetViewPort().GetWidth()==Config::Instance().GetInt("DEFAULT_VIEWPORT_WIDTH"), s.str().c_str());
+		AssertMsg_Test(renderer.GetViewPort().GetWidth()==(size_t)Config::Instance().GetInt("DEFAULT_VIEWPORT_WIDTH"), s.str().c_str());
 	}
 
 #if !defined(ES1) && !defined(ES2)
